Renamed common_defines.h to saix_common_defines.h
